truoc khi pull code xu ly frontend 6/3/2026

+++ backend/seal-hackathon-system/src/main/java/com/minhtung/hackathon/entity/Team.java
@@ -80,7 +80,7 @@ public class Team {
         return leaderID;
     }
 
-    public void setLeaderID(int leaderID) {
+    public void setLeaderID(long leaderID) {
         this.leaderID = leaderID;
     }
